+++ src/containers/OrderReviewPage.jsx
@@ -22,7 +22,7 @@
 import React from 'react';
 import ReactRouterPropTypes from 'react-router-prop-types';
 import intl from 'react-intl-universal';
-import { login } from '../utils/AuthService';
+import { fetchOrderData, submitOrder } from '../utils/AuthService';
 import {
   isAnalyticsConfigured, trackAddItemAnalytics, trackAddTransactionAnalytics, sendAnalytics,
 } from '../utils/Analytics';
@@ -31,40 +31,9 @@ import ShippingOptionContainer from '../components/shippingoption.container';
 import AddressContainer from '../components/address.container';
 import CheckoutSummaryList from '../components/checkout.summarylist';
 import OrderTableMain from '../components/ordertable.main';
-import cortexFetch from '../utils/Cortex';
 import './OrderReviewPage.less';
 
 const Config = require('Config');
-
-const zoomArray = [
-  // zooms for checkout summary
-  'defaultcart',
-  'defaultcart:total',
-  'defaultcart:discount',
-  'defaultcart:order',
-  'defaultcart:order:tax',
-  'defaultcart:order:total',
-  'defaultcart:appliedpromotions:element',
-  // zoom for billing address
-  'defaultcart:order:billingaddressinfo:billingaddress',
-  // zoom for shipping address
-  'defaultcart:order:deliveries:element:destinationinfo:destination',
-  // zoom for shipping option
-  'defaultcart:order:deliveries:element:shippingoptioninfo:shippingoption',
-  // zoom for payment method
-  'defaultcart:order:paymentmethodinfo:paymentmethod',
-  // zooms for table items
-  'defaultcart:lineitems:element',
-  'defaultcart:lineitems:element:total',
-  'defaultcart:lineitems:element:item',
-  'defaultcart:lineitems:element:item:code',
-  'defaultcart:lineitems:element:item:price',
-  'defaultcart:lineitems:element:item:definition',
-  'defaultcart:lineitems:element:item:definition:options:element',
-  'defaultcart:lineitems:element:item:definition:options:element:value',
-  // zoom for purchaseform
-  'defaultcart:order:purchaseform',
-];
 
 class OrderReviewPage extends React.Component {
   static propTypes = {
@@ -80,70 +49,41 @@ class OrderReviewPage extends React.Component {
   }
 
   componentDidMount() {
-    this.fetchOrderData();
+    this.refreshOrderData();
   }
 
-  fetchOrderData() {
-    login().then(() => {
-      cortexFetch(`/?zoom=${zoomArray.sort().join()}`,
-        {
-          headers: {
-            'Content-Type': 'application/json',
-            Authorization: localStorage.getItem(`${Config.cortexApi.scope}_oAuthToken`),
-          },
-        })
-        .then(res => res.json())
-        .then((res) => {
-          this.setState({
-            orderData: res._defaultcart[0],
-          });
-        })
-        .catch((error) => {
-          // eslint-disable-next-line no-console
-          console.error(error.message);
+  refreshOrderData() {
+    fetchOrderData()
+      .then((res) => {
+        this.setState({
+          orderData: res._defaultcart[0],
         });
-    });
+      })
+      .catch((error) => {
+        // eslint-disable-next-line no-console
+        console.error(error.message);
+      });
   }
 
   completeOrder() {
     this.setState({
       isLoading: true,
     });
-    const purchaseZoomArray = [
-      'paymentmeans:element',
-      'shipments:element:destination',
-      'shipments:element:shippingoption',
-      'billingaddress',
-      'discount',
-      'appliedpromotions:element',
-      'lineitems:element',
-      'lineitems:element:options:element',
-      'lineitems:element:options:element:value',
-    ];
     const { orderData } = this.state;
     const { history } = this.props;
     const purchaseform = orderData._order[0]._purchaseform[0].links.find(link => link.rel === 'submitorderaction').uri;
-    login().then(() => {
-      cortexFetch(`${purchaseform}?followlocation&zoom=${purchaseZoomArray.sort().join()}`, {
-        method: 'post',
-        headers: {
-          'Content-Type': 'application/json',
-          Authorization: localStorage.getItem(`${Config.cortexApi.scope}_oAuthToken`),
-        },
-      })
-        .then(res => res.json())
-        .then((res) => {
-          this.setState({
-            isLoading: false,
-          });
-          this.trackTransactionAnalytics();
-          history.push('/purchaseReceipt', { data: res });
-        })
-        .catch((error) => {
-          // eslint-disable-next-line no-console
-          console.error(error.message);
+    submitOrder(purchaseform)
+      .then((res) => {
+        this.setState({
+          isLoading: false,
         });
-    });
+        this.trackTransactionAnalytics();
+        history.push('/purchaseReceipt', { data: res });
+      })
+      .catch((error) => {
+        // eslint-disable-next-line no-console
+        console.error(error.message);
+      });
   }
 
   trackTransactionAnalytics() {
